BoilerList/johnslist/requirements.txt
Django==1.10.6
django-filter==1.0.1
django-guardian==1.4.6
django-model-utils==2.4
django-notifications-hq==1.0
django-sniplates==0.4.1
django-widget-tweaks==1.4.1
easy-thumbnails==2.3
jsonfield==1.0.3
Pillow==3.1.0
pytz==2015.7
six==1.10.0
django-admin-rangefilter==0.1.3
django-import-export==0.5.1
django-debug-toolbar==1.6
django-password-reset==0.9
